remove /doc and /exportQuarto chat commands
