+++ extensions/positron-assistant/package.nls.json
@@ -9,9 +9,7 @@
 	"chatParticipants.fullName": "Positron Assistant",
 	"chatParticipants.ask.description": "Ask Assistant",
 	"chatParticipants.agent.description": "Complete tasks in your workspace",
-	"chatParticipants.commands.doc.description": "Generate documentation comments in the active editor",
 	"chatParticipants.commands.explain.description": "Explain the code in the active editor",
-	"chatParticipants.commands.exportQuarto.description": "Export the conversation so far into a new Quarto document.",
 	"chatParticipants.commands.fix.description": "Fix issues in the active editor or Console",
 	"chatParticipants.edit.description": "Edit files in your workspace",
 	"configuration.title": "Positron Assistant",
