--- extensions/positron-assistant/src/commands/doc.ts
@@ -1,29 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (C) 2025 Posit Software, PBC. All rights reserved.
- *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import * as vscode from 'vscode';
-
-import { PositronAssistantChatContext } from '../participants.js';
-import { PromptRenderer } from '../promptRender.js';
-
-export const DOC_COMMAND = 'doc';
-
-/**
- * Handler for the custom chat participant command `/doc`.
- */
-export async function docHandler(
-	_request: vscode.ChatRequest,
-	context: PositronAssistantChatContext,
-	response: vscode.ChatResponseStream,
-	_token: vscode.CancellationToken,
-	handleDefault: () => Promise<vscode.ChatResult | void>
-) {
-	response.progress(vscode.l10n.t('Generating documentation...'));
-
-	const prompt = PromptRenderer.renderCommandPrompt(DOC_COMMAND, _request).content;
-	context.systemPrompt += `\n\n${prompt}`;
-
-	return handleDefault();
-}
