+++ extensions/positron-assistant/src/commands/index.ts
@@ -3,16 +3,13 @@
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as vscode from 'vscode';
 import * as positron from 'positron';
 import { FIX_COMMAND, fixHandler } from './fix.js';
-import { EXPORT_QUARTO_COMMAND, quartoHandler } from './quarto.js';
 import { EXPLAIN_COMMAND, explainHandler } from './explain.js';
-import { DOC_COMMAND, docHandler } from './doc.js';
 import { log } from '../log.js';
 import {
+	IChatRequestHandler,
 	PositronAssistantAgentParticipant,
-	PositronAssistantChatContext,
 	PositronAssistantChatParticipant,
 	PositronAssistantEditorParticipant,
 	PositronAssistantEditParticipant,
@@ -20,25 +17,6 @@ import {
 	PositronAssistantTerminalParticipant
 } from '../participants.js';
 import { PromptMetadata, PromptMetadataMode, PromptRenderer } from '../promptRender.js';
-
-/**
- * A function that handles chat requests.
- *
- * @param request The chat request to handle.
- * @param context The chat context for the request.
- * @param response The response stream for the request.
- * @param token A cancellation token for the request.
- * @param handleDefault A function to call the default request handler.
- */
-export interface IChatRequestHandler {
-	(
-		request: vscode.ChatRequest,
-		context: PositronAssistantChatContext,
-		response: vscode.ChatResponseStream,
-		token: vscode.CancellationToken,
-		handleDefault: () => Promise<vscode.ChatResult | void>
-	): Promise<vscode.ChatResult | void>;
-}
 
 function registerAssistantCommand(command: string, handler: IChatRequestHandler) {
 	let metadata: PromptMetadata<PromptMetadataMode[]>;
@@ -80,8 +58,6 @@ function registerAssistantCommand(command: string, handler: IChatRequestHandler)
 }
 
 export function registerAssistantCommands() {
-	registerAssistantCommand(DOC_COMMAND, docHandler);
 	registerAssistantCommand(FIX_COMMAND, fixHandler);
 	registerAssistantCommand(EXPLAIN_COMMAND, explainHandler);
-	registerAssistantCommand(EXPORT_QUARTO_COMMAND, quartoHandler);
 }
